fixed correct benching of hard, semi, soft

+++ bench_RD/timer.h
@@ -45,7 +45,7 @@ uint64_t end_timer(void)
 
 int compare_u64(const void *x, const void *y)
 {
-  double xx = *(uint64_t*)x, yy = *(uint64_t*)y;
+  uint64_t xx = *(uint64_t*)x, yy = *(uint64_t*)y;
   if (xx < yy) return -1;
   if (xx > yy) return  1;
   return 0;
